Remove redundant create-connection option in Call REST API activity form
The Connection field rendered two create actions from two components:
the generic 'Create New Http' link (NodeReferenceSelectEditor, driven
by searchNodesKind) and the LS-metadata-driven 'Add new HTTP
connection' button (ConnectionEditor, metadata.connectors).

Make NodeReferenceSelectEditor the single owner of the create
affordance: it renders the explicit connector actions when the field
carries metadata.connectors, otherwise the generic link - never both.
ConnectionEditor reduces to label + form registration + the reference
select.

Fixes https://github.com/wso2/product-integrator/issues/2064
